+++ crates/fluvio-test/src/tests/mod.rs
@@ -1,3 +1,4 @@
 pub mod smoke;
 pub mod concurrent;
+pub mod longevity;
 pub mod producer_stress;