+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes
 
 ## Platform Version 0.9.6 - UNRELEASED
+* Improve display representation for some variants in FluvioError type ([#1581](https://github.com/infinyon/fluvio/issues/1581))
+* Add spinner to `fluvio cluster --local --setup` command ([1574](https://github.com/infinyon/fluvio/pull/1574))
+* Add `--format` string for custom Consumer printouts ([#1593](https://github.com/infinyon/fluvio/issues/1593))
 
 ## Platform Version 0.9.5 - 2021-09-02
 * Update `Debug` printout for `SmartStreamWasm` to reduce noise ([#1524](https://github.com/infinyon/fluvio/pull/1524))